Cleanup Program.cs
