+++ Src/NationsConverterWeb/NationsConverterWeb/Configuration/CacheConfiguration.cs
@@ -1,0 +1,9 @@
+﻿namespace NationsConverterWeb.Configuration;
+
+public static class CacheConfiguration
+{
+    public static void AddCacheServices(this IServiceCollection services)
+    {
+        services.AddHybridCache();
+    }
+}
