+++ Src/NationsConverterWeb/NationsConverterWeb/Program.cs
@@ -1,300 +1,35 @@
-using AspNet.Security.OAuth.Discord;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.CookiePolicy;
-using Microsoft.AspNetCore.HttpOverrides;
-using Microsoft.AspNetCore.ResponseCompression;
-using Microsoft.AspNetCore.StaticFiles;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.FileProviders;
-using NationsConverterWeb;
-using NationsConverterWeb.Authentication;
-using NationsConverterWeb.Components;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Logs;
-using OpenTelemetry.Trace;
-using Microsoft.Extensions.Caching.Hybrid;
-using NationsConverterWeb.BulkFixers;
-using Serilog;
-using Serilog.Sinks.SystemConsole.Themes;
-using System;
+using NationsConverterWeb.Configuration;
 
 GBX.NET.Gbx.LZO = new GBX.NET.LZO.MiniLZO();
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme)
-    .AddCookie(options =>
-    {
-        options.LoginPath = "/";
-        options.AccessDeniedPath = "/";
-    })
-    .AddDiscord(options =>
-    {
-        options.ClientId = builder.Configuration["Discord:ClientId"] ?? throw new InvalidOperationException("Discord ClientId is missing");
-        options.ClientSecret = builder.Configuration["Discord:ClientSecret"] ?? throw new InvalidOperationException("Discord ClientSecret is missing");
-        options.AccessDeniedPath = "/";
-        options.ClaimActions.MapJsonKey(DiscordAdditionalClaims.GlobalName, "global_name");
-
-        options.Events.OnCreatingTicket = DiscordAuthenticationTicket.OnCreatingTicketAsync;
-    });
-
-builder.Services.AddCascadingAuthenticationState();
-
-builder.Services.AddDirectoryBrowser();
-
-builder.Services.AddResponseCompression(options =>
-{
-    options.EnableForHttps = true;
-    options.Providers.Add<BrotliCompressionProvider>();
-    options.Providers.Add<GzipCompressionProvider>();
-});
-
-builder.Services.Configure<ForwardedHeadersOptions>(options =>
-{
-    options.ForwardedHeaders =
-        ForwardedHeaders.XForwardedFor |
-        ForwardedHeaders.XForwardedProto;
-    options.KnownNetworks.Clear();
-    options.KnownProxies.Clear();
-});
-
-builder.Services.AddDbContextFactory<AppDbContext>(options =>
-{
-    var connectionStr = builder.Configuration.GetConnectionString("DefaultConnection");
-    options.UseMySql(connectionStr, ServerVersion.AutoDetect(connectionStr));
-});
-
-builder.Services.AddTransient<IClaimsTransformation, ClaimsTransformation>();
-
-builder.Services.AddScoped<RevertMaterialPhysicsBulkFixer>();
-builder.Services.AddScoped<CheckpointTerrainModifierBulkFixer>();
-
-// Add services to the container.
-builder.Services.AddRazorComponents()
-    .AddInteractiveServerComponents()
-    .AddInteractiveWebAssemblyComponents();
-
-Log.Logger = new LoggerConfiguration()
-    .ReadFrom.Configuration(builder.Configuration)
-    .Enrich.FromLogContext()
-    .WriteTo.Console(theme: AnsiConsoleTheme.Sixteen, applyThemeToRedirectedOutput: true)
-    .WriteTo.OpenTelemetry(options =>
-    {
-        options.Endpoint = builder.Configuration["OTEL_EXPORTER_OTLP_ENDPOINT"];
-        options.Protocol = builder.Configuration["OTEL_EXPORTER_OTLP_PROTOCOL"]?.ToLowerInvariant() switch
-        {
-            "grpc" => Serilog.Sinks.OpenTelemetry.OtlpProtocol.Grpc,
-            "http/protobuf" or null or "" => Serilog.Sinks.OpenTelemetry.OtlpProtocol.HttpProtobuf,
-            _ => throw new NotSupportedException($"OTLP protocol {builder.Configuration["OTEL_EXPORTER_OTLP_PROTOCOL"]} is not supported")
-        };
-        options.Headers = builder.Configuration["OTEL_EXPORTER_OTLP_HEADERS"]?
-            .Split(',', StringSplitOptions.RemoveEmptyEntries)
-            .Select(x => x.Split('=', 2, StringSplitOptions.RemoveEmptyEntries))
-            .ToDictionary(x => x[0], x => x[1]) ?? [];
-    })
-    .CreateLogger();
-
-builder.Services.AddSerilog();
-
-builder.Services.AddOpenTelemetry()
-    .WithMetrics(options =>
-    {
-        options
-            .AddAspNetCoreInstrumentation()
-            .AddHttpClientInstrumentation()
-            .AddRuntimeInstrumentation()
-            .AddProcessInstrumentation()
-            .AddOtlpExporter();
-
-        options.AddMeter("System.Net.Http");
-    })
-    .WithTracing(options =>
-    {
-        if (builder.Environment.IsDevelopment())
-        {
-            options.SetSampler<AlwaysOnSampler>();
-        }
-
-        options
-            .AddAspNetCoreInstrumentation()
-            .AddHttpClientInstrumentation()
-            .AddEntityFrameworkCoreInstrumentation()
-            .AddOtlpExporter();
-    });
-builder.Services.AddMetrics();
-
-builder.Services.AddHybridCache();
+builder.Services.AddDomainServices();
+builder.Services.AddWebServices(builder.Configuration);
+builder.Services.AddCacheServices();
+builder.Services.AddDataServices(builder.Configuration);
+builder.Services.AddTelemetryServices(builder.Configuration, builder.Environment);
 
 var app = builder.Build();
 
-using (var scope = app.Services.CreateScope())
-{
-    var dbContext = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-
-    if (dbContext.Database.IsRelational())
-    {
-        dbContext.Database.Migrate();
-    }
-}
-
-app.UseForwardedHeaders();
+app.MigrateDatabase();
 
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
 {
     app.UseWebAssemblyDebugging();
+    app.UseForwardedHeaders();
 }
 else
 {
     app.UseExceptionHandler("/Error", createScopeForErrors: true);
+    app.UseForwardedHeaders();
     // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
 
-app.UseStaticFiles(new StaticFileOptions()
-{
-    ContentTypeProvider = new FileExtensionContentTypeProvider
-    {
-        Mappings = { [".mux"] = "application/octet-stream" }
-    }
-});
-
-app.UseCookiePolicy(new CookiePolicyOptions
-{
-    MinimumSameSitePolicy = SameSiteMode.Lax,
-    Secure = CookieSecurePolicy.Always,
-    HttpOnly = HttpOnlyPolicy.Always
-});
-
-app.UseAuthentication();
-app.UseAuthorization();
-
-var dataDir = Path.Combine(AppContext.BaseDirectory, "Data");
-Directory.CreateDirectory(dataDir);
-
-/*app.UseStaticFiles(new StaticFileOptions
-{
-    FileProvider = new PhysicalFileProvider(dataDir),
-    RequestPath = "/data",
-    ServeUnknownFileTypes = true
-});*/
-
-app.UseDirectoryBrowser(new DirectoryBrowserOptions
-{
-    FileProvider = new PhysicalFileProvider(dataDir),
-    RequestPath = "/data"
-});
-
-if (!app.Environment.IsDevelopment())
-{
-    app.UseResponseCompression();
-}
-
-app.MapGet("/login-discord", async (HttpContext context) =>
-{
-    await context.ChallengeAsync(DiscordAuthenticationDefaults.AuthenticationScheme, new()
-    {
-        RedirectUri = "/dashboard"
-    });
-}).AllowAnonymous();
-
-app.MapGet("/logout", async (HttpContext context) =>
-{
-    await context.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
-    context.Response.Redirect("/");
-});
-
-app.MapGet("/blockicon/{name}", async (
-    HttpContext context,
-    AppDbContext db,
-    IWebHostEnvironment env,
-    HybridCache cache,
-    string? name) =>
-{
-    if (name is null)
-    {
-        context.Response.StatusCode = StatusCodes.Status404NotFound;
-        return;
-    }
-
-    try
-    {
-        var block = await cache.GetOrCreateAsync($"blockicon_{name}", async token =>
-        {
-            return await db.Blocks
-                .Select(x => new { x.Name, x.IconWebp, x.CreatedAt })
-                .AsNoTracking()
-                .FirstOrDefaultAsync(x => x.Name == name, token);
-        }, new() { Expiration = TimeSpan.FromMinutes(5) }, cancellationToken: context.RequestAborted);
-
-        if (block is null)
-        {
-            context.Response.StatusCode = StatusCodes.Status404NotFound;
-            return;
-        }
-
-        var lastModified = string.IsNullOrWhiteSpace(block.IconWebp)
-            ? env.WebRootFileProvider.GetFileInfo(Path.Combine("img, bloc.webp")).LastModified
-            : block.CreatedAt;
-
-        var eTag = $"\"{lastModified.Ticks}\"";
-
-        if (context.Request.Headers.TryGetValue("If-None-Match", out var requestEtag) && requestEtag == eTag)
-        {
-            context.Response.StatusCode = StatusCodes.Status304NotModified;
-            return;
-        }
-
-        if (context.Request.Headers.TryGetValue("If-Modified-Since", out var ifModifiedSince))
-        {
-            if (DateTime.TryParse(ifModifiedSince, out var modifiedSince) && modifiedSince >= lastModified)
-            {
-                context.Response.StatusCode = StatusCodes.Status304NotModified;
-                return;
-            }
-        }
-
-        context.Response.ContentType = "image/webp";
-        context.Response.Headers.ETag = eTag;
-        context.Response.Headers.LastModified = lastModified.ToString("R");
-
-        if (string.IsNullOrWhiteSpace(block.IconWebp))
-        {
-            await context.Response.SendFileAsync(Path.Combine(env.WebRootPath, "img", "bloc.webp"), context.RequestAborted);
-            return;
-        }
-
-        var iconBytes = Convert.FromBase64String(block.IconWebp);
-
-        await context.Response.Body.WriteAsync(iconBytes, context.RequestAborted);
-    }
-    catch (OperationCanceledException)
-    {
-        context.Response.StatusCode = StatusCodes.Status500InternalServerError;
-    }
-});
-
-app.MapGet("/assets/{name}", async (string name, AppDbContext db, CancellationToken cancellationToken) =>
-{
-    var release = await db.AssetReleases
-        .AsNoTracking()
-        .OrderByDescending(x => x.ReleasedAt)
-        .FirstOrDefaultAsync(x => x.Name == name, cancellationToken);
-
-    if (release is null)
-    {
-        return Results.NotFound();
-    }
-
-    return Results.File(release.Data, "application/zip", $"{release.Name}.nc2", lastModified: release.ReleasedAt);
-});
-
-app.UseAntiforgery();
-
-app.MapRazorComponents<App>()
-    .AddInteractiveServerRenderMode()
-    .AddInteractiveWebAssemblyRenderMode()
-    .AddAdditionalAssemblies(typeof(NationsConverterWeb.Client._Imports).Assembly);
+app.UseAuthMiddleware();
+app.UseSecurityMiddleware();
+app.UseEndpointMiddleware();
 
 app.Run();
